Moved offcanvas component for courseList to a different file.
